[DO NOT MERGE] Additional socket hooks for netd
* recvfrom probably not needed

Allow netd to hook bind(), getsockname(), getpeername(), and
recvfrom(). This allows further adjustment or blocking of socket
behavior, such as the binding of abstract sockets that are typically
visible throughout the system via a shared namespare.

This commit is intended to aid in averting the type of communication
possible via abstract socket name checking, as demonstrated by e.g.
LocalSocketChat.

Issue: calyxos#1180
Change-Id: Ie7cd485a5ded085834fa573fc6337073500b183c