setjmp (x86/x86-64): checksum the whole buffer.
arm64 (and its copy-and-paste riscv64) both cover all the other [used]
fields in their checksum, but for some reason x86/x86-64 don't, which
seems like a bug.

I'm really here because of the "do we need a cookie _and_ a checksum?"
question, but while I'm still thinking about that, let's at least make
x86/x86-64 the same as arm64/riscv64.

(arm32 has a similar bug, but I'll fix that in a separate change.)

Change-Id: Icb09e869bbca969e3463b310a0b944c7c9515a15