Reland "setjmp (x86/x86-64): use cmov."
This reverts commit 6fec578c7823cf6fa726e564fee57bb8d02a5bed, and
switches from ebx to edx so we don't corrupt ebx.

The unit tests can't tell the difference, because we only check the
values in the floating-point callee-saves. I could have sworn that we
used to have tests for the integer callee-saves and had removed them
because they'd caused problems, but the git history disagrees.

I'm trying to write the missing test, but struggling to get the compiler
itself not to use any of the registers. In particular, on x86-64 ebx is
the first register it takes for its own use. I'll keep trying, but will
merge this first.

Change-Id: I6a8aaa774c484f1fbb2c46f44e4b4d2772c4bac7