icu4x: Remove dependency on std
Currently, icu4x implicitly depends on libstd, but most of it gets DCE'd
out. Unfortunately, this introduces a single thread local variable in
icu4x_bionic, which gets linked into the linker. Since the linker is
unable to relocate TLS entries in itself, this prevents us from enabling
ELF TLS.

This change rewires the inclusion of icu4x to select the no_std variant
and hook panic and alloc up to bionic.

Bug: 446969250
Test: Boot cuttlefish with ELF TLS enabled
Change-Id: I75b0907cb32f244fb4fbe1076748ae9907497e41