tz: microoptimize strlcpy() call.
According to our profiling data, __bionic_get_system_tz() is our most frequent caller to strlcpy() ... completely unnecessarily. The "bargain" is basically that the buffer needs to be PROP_VALUE_MAX bytes, but that's not something that (a) is likely to change and (b) is something that use of strlcpy() really helps with. If this were a public API there would be more call for the pointer+size_t but given that it's solely internal, this seems like we're just wasting everyone's time.

We also don't need to zero-initialize the buffer.

None of this probably matters much, because it's still not exactly a hotspot, but it _is_ the most visible strlcpy() call site, so while we're here, let's just reduce the waste.

Change-Id: Ic41d051b932989e777850ced022e670da2113918