Add a trivial memchr()/memcpy()-based memccpy().
I'll be along shortly with arm64/x86-64 assembler from FreeBSD, but for
all the other architectures we can still easily beat the llvm-libc
implementation.

This one's a little bit interesting because it is 2x slower for very
small n, but it's 2x faster by 32 bytes, and 10x faster by 512 bytes. We
don't have many existing call sites for memccpy(), but it doesn't seem
unreasonable to assume that it's mostly useful for large n?

Realistically any architecture we care about should move to assembler
anyway...

Change-Id: I08e784a1e7c0b645b1effeb16d4f3eda7f55e66a