libc: naively migrate portable-simd to highway
This ended up being much less effort than I thought, thanks in part to
highway supporting builds without libcxx, and in part due to header-only
efforts that are underway upstream.

The commit message says 'naively' because this is a direct translation
of existing code to highway. Now that we're using highway, there may be
better idioms to use, but those can be swapped to more incrementally.

Bug: 453997143
Test: Examined the `objdump` of v2 and v3 strlens before and after.
Test: v2 are identical; v3 is identical except that the old code has a
Test: `cltq` on `zero_mask` in one path and hwy doesn't. `cltq` is
Test: sign-extension, and since we only care about the lowest set bit in
Test: the mask, it's useless.

Change-Id: I8f43ab3efee3b21fd1a4c3588519d81dbeba2a21