libc: add portable-simd memchr
Brya benchmarking doc says that this is strictly faster than our current
verison for both SSE and AVX2 (AVX2 is ~1.4x faster in large cases).

Methodology was somewhat primitive like strlen; my results are here:
https://docs.google.com/document/d/1duN9BCkviZ0IfrxgMDRMdMgr4ESS012wW09Bq-zxCEQ/edit?tab=t.0

As the comments allude to, I tried different numbers of iterations
within the 'large loop' case. The doc presents the case that's in this
CL.

Bug: 452714218
Test: Benchmarked on brya

Change-Id: I0fca301dc2577b860ec4670f9e9d97c2afd5d07a