libc: add psimd wcslen
Clang will optimize some loops to this, and it's easy to model on the
existing `strlen` implementation, so do that.

If this becomes too much of a pain to maintain on top of strlen, we can
always write a simpler ~30LOC version. Either should do better than the
current byte-by-byte impl we're pulling in for all platforms.

Bug: 452714218
Test: Unittests, go/bionic-psimd-benchmarks?tab=t.t2kk09rvng7i
Change-Id: I5d42ccbe51d01d65dd8abcf6080fd247a5d03d07