libc: memchr isn't guaranteed its entire buffer
Chromium is crashing, seemingly because `strstr` calls `memchr` with a
too-long size. This is considered well-defined since C11:
b/460498586#comment4

This is a partial reland of strnlen
(I8ed0dc3b3d97b794d0849badc4a2abe8a4658e90), which had logic for "the
size is an upper-bound, not a guarantee," which is what we need here.

Bug: 460498586, 452714218
Test: mma && ./tests/run-on-host.sh 64 --gtest_filter='string.memchr*'
Change-Id: I1cdfc1bcc81eba7e76b517b1b6d21083a971d1d1