Add a relative pc and shared library to trace.
Rewrite the code to still support the fdtrack_iterate function,
but add an internal version that is used by both. One only uses
the function name and function offsets. The other uses the unwinder
frame data directly, and the default one will also display
the relative pc and the shared library in the backtrace.

Other fixes:
- Always demangle the function names.
- Add a lock around the fatal error to prevent mutliple fatal
  happening and accessing the static variables at the same tie.

Bug: 462283002

Test: All unit tests pass.
Test: Run setprop persist.sys.debug.fdtrack_enable_threshold 1
Test: and kill -39 <system_server>
Test: Verify the fdtrack output.
Test: Did the same, but triggered the fatal path.
Change-Id: I10d31088e6f96f624b62af8ee384f34d86a67ca1